// src/utils/menu-to-grid/type.ts
export type DOMType = {
   el: any
   title: any
   grid: any
   images: any
}

export type PreviewItemType = {
	DOM: DOMType
}

export type RowDomType = {
   el: any
   title: any
   titleWrap: any
   imagesWrap: any
   images: any
}

export type RowType = {
   DOM: RowDomType
   previewItem: any
}